Uniform pilot-agent and pilot-discover 's code style
