+++ pilot/cmd/pilot-agent/app/cmd_test.go
@@ -11,14 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-package main
+package app
 
 import (
 	"testing"
 
 	"github.com/onsi/gomega"
-
 	"istio.io/istio/pilot/pkg/model"
 )
